wiki.py: complete overhaul
Since all embeds here lead to the wiki, repeating it doesn't make sense.
embed is now a method used to create an embed with the author and url predefined.
self.embed("title") where title is the name of the embed.
embed.url should now be appended with the URL path after "https://wiki.ds-homebrew.com/"

yay for no hardcoding