Move `CustomView` interaction_check to ViewPages class instead
There are now more commands using the CustomView class, and for those
commands, having the interaction check deny button access to other
users is not ideal, whereas in the initial set of API commands, it was.

Let's just let the downstream classes handle interaction checks.

Also, copy-paste the NetInfoManager ephemeral message to ViewPages, as
the intention is identical anyway.